+++ packages/workspace-core/src/lib/types/index.ts
@@ -1,1 +1,3 @@
 export * from './types';
+export * from './callback';
+export * from './events';
